Replace writecrops with more comprehensive scripts
Add a new script, writeedits, that outputs a file with various edits,
which can then be "used" via a simple bash script.

Edits are output in a TSV, starting with the filename, then a
ffmpeg-friendly filter chain, then the start time, and then the end
time.

There is no deduping done to output of the files; so any scripts should
run the appropriate deduplication routines