[MPV] Add negative crop, change autocrop's filter order
Some videos are padded with white, for some godawful reason. The default
autocrop will miss this padding, as it looks for black padding.
Inverting/negating the image color lets it happily crop these areas.

Unfortunately, the original autocrop script prepends its crop filter to
the filtergraph, meaning it misses the negation. Changing it to just
append solves this.