[MPV] Fix rotate function in writeedits.js
Make the rotate function in writeedits use transpose or flip instead of the
rotate filter.

By default, the rotate filter has a cropping effect, as it refuses to
exceed the normal bounds of the original image. You can override this
with output width and height params, but the trigonometry for resizing
the image to fit within these domains is up to you.

Most uses of the rotate function are to correct image that is shot in
vertical video. So we can use a mix of transpose and flip filters to
solve this.

Unfortunately, this means arbitrary rotations no longer work, just 90
degree increments. If you want arbitrary rotation, do it by hand