[surfingkeys] disable neovim
Neovim in surfingkeys has NEVER worked for me. Every time i'd try to
invoke the built-in neovim editor, it would freeze the entire
surfingkeys process for that tab, and I'd have to reload the page

I opened a ticket about this (Surfingkeys#1619), but no fix has arrived
in the 6 months since, so I'm disabling it for now.

I use Firenvim anyways, for large edits, so this is mostly just about
being able to use the ace editor for URLs, Selects, and single-line
inputs