[🐟] Direnv loads need to happen after path is set
Direnv is typically installed by homebrew, which has to be added to the
path. Running the direnv loader as a fish/conf.d file is run _before_
path was set, before config.fish is loaded, and was failing.

I could hardcode paths, or use brew prefix, but its easier to just put
it at the end of the config file